secure by default
finally got Commodo signed certs to work - the trick (for now) is to not include the entire signing chain in the self.pem file in noVNC - you only want the private key and the CA signed cert. This is a problem because Firefox wants the whole chain, but that is something to fix later.

Changed noVNC/launch.sh to require https connections

Signed-off-by: Mark McCahill <mark.mccahill@duke.edu>